remove tradingview chart and rename
