+++ src/components/Chart/index.tsx
@@ -1,13 +1,11 @@
 import React from "react";
-// import { Chart as BaseChart } from "./Chart";
-import { TradeChart as BaseChart1 } from "./Chart1";
+import { TradeChart as BaseChart } from "./Chart1";
 import { ChartErrorBoundary } from "./ChartErrorBoundary";
 
-const ChartWithErrorBoundary = (props: React.ComponentProps<typeof BaseChart1>) => {
+const ChartWithErrorBoundary = (props: React.ComponentProps<typeof BaseChart>) => {
   return (
     <ChartErrorBoundary>
-      {/* <BaseChart1 {...props} /> */}
-      <BaseChart1 {...props} />
+      <BaseChart {...props} />
     </ChartErrorBoundary>
   );
 };
